Update existing files to support the new mode
